# apps/api/spec/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "acting-api",
    "description" : "연기 분석 플랫폼 API v2. Bearer access token을 사용합니다.",
    "version" : "0.1.0"
  },
  "paths" : {
...
    "/v2/consents" : {
      "post" : {
        "tags" : [ "v2-consents" ],
        "summary" : "Record Consent",
        "operationId" : "record_consent_v2_consents_post",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConsentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "새 결정을 쌓았다",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentEventResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "지금 결정과 같은 결정이다. 행을 늘리지 않고 이미 있는 결정을 돌려준다",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentEventResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "현재 판이 아닌 문서에 결정을 보냈다",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentDocumentOutdatedError"
                }
              }
            }
          },
          "422" : {
            "description" : "Validation Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security" : [ {
          "HTTPBearer" : [ ]
        } ]
      }
    },
...
    "/v2/consents/pending" : {
      "get" : {
        "tags" : [ "v2-consents" ],
        "summary" : "List Pending Documents",
        "operationId" : "list_pending_documents_v2_consents_pending_get",
        "responses" : {
          "200" : {
            "description" : "Successful Response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentDocumentsResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "HTTPBearer" : [ ]
        } ]
      }
    },
...
    "/v2/consents/entry" : {
      "get" : {
        "tags" : [ "v2-consents" ],
        "summary" : "Get Consent Entry",
        "operationId" : "get_consent_entry_v2_consents_entry_get",
        "responses" : {
          "200" : {
            "description" : "Successful Response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentEntryResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "HTTPBearer" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ConsentDecision" : {
        "type" : "string",
        "enum" : [ "granted", "declined" ],
        "title" : "ConsentDecision"
      },
      "ConsentRequest" : {
        "type" : "object",
        "properties" : {
          "document_id" : {
            "type" : "string",
            "title" : "Document Id"
          },
          "action" : {
            "$ref" : "#/components/schemas/ConsentDecision"
          },
          "age_confirmed" : {
            "type" : "boolean",
            "title" : "Age Confirmed"
          }
        },
        "required" : [ "action", "document_id" ],
        "title" : "ConsentRequest"
      },
      "ConsentAction" : {
        "type" : "string",
        "enum" : [ "granted", "declined", "revoked" ],
        "title" : "ConsentAction"
      },
      "ConsentEventResponse" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Id"
          },
          "document_id" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Document Id"
          },
          "action" : {
            "$ref" : "#/components/schemas/ConsentAction"
          },
          "occurred_at" : {
            "type" : "string",
            "format" : "date-time",
            "title" : "Occurred At"
          }
        },
        "required" : [ "action", "document_id", "id", "occurred_at" ],
        "title" : "ConsentEventResponse"
      },
      "ConsentDocumentOutdatedError" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "detail" : {
            "type" : "string",
            "enum" : [ "consent_document_outdated" ],
            "title" : "Detail"
          }
        },
        "required" : [ "detail" ],
        "title" : "ConsentDocumentOutdatedError"
      },
...
      "ConsentDocument" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Id"
          },
          "type" : {
            "$ref" : "#/components/schemas/ConsentType"
          },
          "version" : {
            "type" : "string",
            "title" : "Version"
          },
          "title" : {
            "type" : "string",
            "title" : "Title"
          },
          "body" : {
            "type" : "string",
            "title" : "Body"
          },
          "required" : {
            "type" : "boolean",
            "title" : "Required"
          },
          "published_at" : {
            "type" : "string",
            "format" : "date-time",
            "title" : "Published At"
          }
        },
        "required" : [ "body", "id", "published_at", "required", "title", "type", "version" ],
        "title" : "ConsentDocument"
      },
      "ConsentType" : {
        "type" : "string",
        "enum" : [ "terms", "privacy", "ai_analysis", "retention" ],
        "title" : "ConsentType"
      },
...
      "HTTPValidationError" : {
        "type" : "object",
        "properties" : {
          "detail" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ValidationError"
            },
            "title" : "Detail"
          }
        },
        "title" : "HTTPValidationError"
      },
      "ValidationError" : {
        "type" : "object",
        "properties" : {
          "loc" : {
            "type" : "array",
            "items" : {
              "anyOf" : [ {
                "type" : "string"
              }, {
                "type" : "integer"
              } ]
            },
            "title" : "Location"
          },
          "msg" : {
            "type" : "string",
            "title" : "Message"
          },
          "type" : {
            "type" : "string",
            "title" : "Error Type"
          },
          "input" : {
            "title" : "Input"
          },
          "ctx" : {
            "type" : "object",
            "title" : "Context"
          }
        },
        "required" : [ "loc", "msg", "type" ],
        "title" : "ValidationError"
      },
...
      "ConsentDocumentsResponse" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "documents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ConsentDocument"
            },
            "title" : "Documents"
          }
        },
        "required" : [ "documents" ],
        "title" : "ConsentDocumentsResponse"
      },
...
      "ConsentEntryDocument" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "title" : "Id"
          },
          "type" : {
            "$ref" : "#/components/schemas/ConsentType"
          },
          "version" : {
            "type" : "string",
            "title" : "Version"
          },
          "title" : {
            "type" : "string",
            "title" : "Title"
          },
          "body" : {
            "type" : "string",
            "title" : "Body"
          },
          "required" : {
            "type" : "boolean",
            "title" : "Required"
          },
          "published_at" : {
            "type" : "string",
            "format" : "date-time",
            "title" : "Published At"
          },
          "current_decision" : {
            "anyOf" : [ {
              "type" : "string",
              "enum" : [ "granted", "declined" ]
            }, {
              "type" : "null"
            } ],
            "title" : "Current Decision"
          },
          "decided_at" : {
            "anyOf" : [ {
              "type" : "string",
              "format" : "date-time"
            }, {
              "type" : "null"
            } ],
            "title" : "Decided At"
          }
        },
        "required" : [ "body", "current_decision", "decided_at", "id", "published_at", "required", "title", "type", "version" ],
        "title" : "ConsentEntryDocument"
      },
      "ConsentEntryResponse" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "entry_status" : {
            "$ref" : "#/components/schemas/ConsentEntryStatus"
          },
          "documents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ConsentEntryDocument"
            },
            "title" : "Documents"
          },
          "undecided_documents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ConsentEntryDocument"
            },
            "title" : "Undecided Documents"
          }
        },
        "required" : [ "documents", "entry_status", "undecided_documents" ],
        "title" : "ConsentEntryResponse"
      },
      "ConsentEntryStatus" : {
        "type" : "string",
        "enum" : [ "allowed", "decision_required" ],
        "title" : "ConsentEntryStatus"
      },
...
    },
    "securitySchemes" : {
      "HTTPBearer" : {
        "type" : "http",
        "scheme" : "bearer"
      }
    }
  }
}